+++ packages/eas-cli/src/utils/statuspageService.ts
@@ -24,6 +24,8 @@ const humanReadableServiceName: Record<StatuspageServiceName, string> = {
   [StatuspageServiceName.EasBuild]: 'EAS Build',
   [StatuspageServiceName.EasSubmit]: 'EAS Submit',
   [StatuspageServiceName.EasUpdate]: 'EAS Update',
+  [StatuspageServiceName.GithubApiRequests]: 'GitHub API Requests',
+  [StatuspageServiceName.GithubWebhooks]: 'Github Webhooks',
 };
 
 function warnAboutServiceOutage(service: StatuspageServiceFragment): void {