+++ packages/eas-cli/src/rollout/actions/NonInteractiveRollout.ts
@@ -1,6 +1,12 @@
 import { EASUpdateAction, EASUpdateContext } from '../../eas-update/utils';
 import { UpdateChannelBasicInfoFragment } from '../../graphql/generated';
-import { ChannelQuery } from '../../graphql/queries/ChannelQuery';
+import {
+  ChannelQuery,
+  UpdateBranchObject,
+  UpdateChannelObject,
+} from '../../graphql/queries/ChannelQuery';
+import { printJsonOnlyOutput } from '../../utils/json';
+import { getRollout, getRolloutInfo, isConstrainedRolloutInfo, isRollout } from '../branch-mapping';
 import {
   CreateRollout,
   NonInteractiveOptions as CreateRolloutNonInteractiveOptions,
@@ -17,6 +23,19 @@ import {
 import { ManageRolloutActions } from './ManageRollout';
 import { MainMenuActions, RolloutActions } from './RolloutMainMenu';
 
+type JSONRolloutOutput = {
+  defaultBranch: UpdateBranchObject;
+  rolledOutBranch: UpdateBranchObject;
+  percentRolledOut: number;
+  runtimeVersion?: string;
+  updatedAt: Date;
+};
+
+type JSONOutput = {
+  hasRollout: boolean;
+  originalRolloutInfo?: JSONRolloutOutput;
+  currentRolloutInfo?: JSONRolloutOutput;
+};
 /**
  * Control a rollout in non interactive mode.
  */
@@ -24,6 +43,7 @@ export class NonInteractiveRollout implements EASUpdateAction<void> {
   constructor(
     private options: {
       channelName?: string;
+      json?: boolean;
       action?: RolloutActions;
     } & Partial<EditRolloutNonInteractiveOptions> &
       Partial<EndRolloutNonInteractiveOptions> &
@@ -47,11 +67,26 @@ export class NonInteractiveRollout implements EASUpdateAction<void> {
       appId: app.projectId,
       channelName,
     });
+    const channelObject = await this.getChannelObjectAsync(ctx, {
+      channelName,
+      runtimeVersion: this.getRuntimeVersion(channelInfo),
+    });
 
     if (!action) {
       throw new Error(`--action is required in non-interactive mode.`);
     }
-    await this.runActionAsync(ctx, action, channelInfo, this.options);
+    const updatedChannelInfo = await this.runActionAsync(ctx, action, channelObject, this.options);
+    const updatedChannelObject = await this.getChannelObjectAsync(ctx, {
+      channelName,
+      runtimeVersion: this.getRuntimeVersion(updatedChannelInfo),
+    });
+    if (this.options.json) {
+      const json = await this.getJsonAsync({
+        originalChannelObject: channelObject,
+        updatedChannelObject,
+      });
+      printJsonOnlyOutput(json);
+    }
   }
 
   async runActionAsync(
@@ -71,5 +106,57 @@ export class NonInteractiveRollout implements EASUpdateAction<void> {
       case ManageRolloutActions.END:
         return await new EndRollout(channelInfo, options).runAsync(ctx);
     }
+  }
+
+  async getJsonAsync({
+    originalChannelObject,
+    updatedChannelObject,
+  }: {
+    originalChannelObject: UpdateChannelObject;
+    updatedChannelObject: UpdateChannelObject;
+  }): Promise<JSONOutput> {
+    return {
+      hasRollout: isRollout(updatedChannelObject),
+      ...(isRollout(originalChannelObject)
+        ? { originalRolloutInfo: await this.getRolloutJsonAsync(originalChannelObject) }
+        : {}),
+      ...(isRollout(updatedChannelObject)
+        ? { currentRolloutInfo: await this.getRolloutJsonAsync(updatedChannelObject) }
+        : {}),
+    };
+  }
+
+  async getRolloutJsonAsync(channelObject: UpdateChannelObject): Promise<JSONRolloutOutput> {
+    const rollout = getRollout(channelObject);
+    return {
+      defaultBranch: rollout.defaultBranch,
+      rolledOutBranch: rollout.rolledOutBranch,
+      percentRolledOut: rollout.percentRolledOut,
+      runtimeVersion: this.getRuntimeVersion(channelObject) ?? undefined,
+      updatedAt: channelObject.updatedAt,
+    };
+  }
+
+  getRuntimeVersion(channelInfo: UpdateChannelBasicInfoFragment): string | undefined {
+    if (isRollout(channelInfo)) {
+      const updatedRolloutInfo = getRolloutInfo(channelInfo);
+      if (isConstrainedRolloutInfo(updatedRolloutInfo)) {
+        return updatedRolloutInfo.runtimeVersion;
+      }
+    }
+    return undefined;
+  }
+
+  async getChannelObjectAsync(
+    ctx: EASUpdateContext,
+    { channelName, runtimeVersion }: { channelName: string; runtimeVersion?: string }
+  ): Promise<UpdateChannelObject> {
+    const { graphqlClient, app } = ctx;
+    const { projectId } = app;
+    return await ChannelQuery.viewUpdateChannelAsync(graphqlClient, {
+      appId: projectId,
+      channelName,
+      ...(runtimeVersion ? { filter: { runtimeVersions: [runtimeVersion] } } : {}),
+    });
   }
 }